+++ src/lib/pipelines/default.js
