+++ src/lib/utils/esri_types.js
