+++ src/templates/service.json
@@ -335,24 +335,7 @@
             "value": "75"
           }
         ]
-      },
-      "labelingInfo" : [
-        {
-          "type": "text",
-          "color": "white",
-          "haloColor": "black",
-          "haloSize": "1px",
-          "text": "You are here",
-          "xoffset": 3,
-          "yoffset": 3,
-          "font": {
-            "size": 12,
-            "family": "Josefin Slab",
-            "weight": "bold"
-          }
-        }
-      ]
-
+      }
     },
     "fields": null,
     "currentVersion": "10.5",
